add stream phase

+++ utils/memory.py
@@ -84,7 +84,7 @@ class OperationalMemory():
         n = samples.shape[0]
         if n >= self.novel_acceptance:
           samples = samples.reshape(samples.shape[0], -1)*255
-          labels = torch.full((n, 1), label, device=self.device, dtype=torch.float) #[200, 1]
+          labels = torch.full((n, 1), label, device=self.device, dtype=torch.int) #[200, 1]
           data = torch.cat((samples, labels), axis=1)
           returned_data_list.append(data)
